+++ test/clj_kondo/test_test.clj
@@ -59,14 +59,9 @@
                      '{:linters {:unresolved-symbol {:level :error}}}))))
 
 (deftest are-test
-  (assert-submaps
-   [{:file "corpus/clojure.test.are.cljc", :row 9, :col 11,
-     :level :warning, :message "unused binding a"}
-    {:file "corpus/clojure.test.are.cljc", :row 9, :col 21,
-     :level :error, :message "Unresolved symbol: z"}]
-   (lint! (io/file "corpus" "clojure.test.are.cljc")
-          {:linters {:unresolved-symbol {:level :error}
-                     :unused-binding {:level :warning}}})))
+  (is (empty? (lint! (io/file "corpus" "clojure.test.are.cljc")
+                     {:linters {:unresolved-symbol {:level :error}
+                                :unused-binding {:level :warning}}}))))
 
 (deftest inline-def-test
   (assert-submaps